requirements.txt
numpy
pandas
torch >=1.9,<2.0
torchaudio >=0.10,<1.0
tqdm
pyyaml
pyannote.audio
scikit-learn
speechbrain